+++ src/models/IntegracaoBsellerErros.js
@@ -3,7 +3,7 @@ const { Model, DataTypes } = require('sequelize');
 class IntegracaoBsellerErros extends Model {
     static init(sequelize) {
         super.init({
-            id_anymarket: DataTypes.INTEGER,
+            id_anymarket: DataTypes.BIGINT,
             fulfillment: DataTypes.BOOLEAN,
             status_anymarket: DataTypes.STRING,
         }, {
