+++ src/models/StatusValidation.js
@@ -3,7 +3,7 @@ const { Model, DataTypes } = require('sequelize');
 class StatusValidation extends Model {
     static init(sequelize) {
         super.init({
-            id_anymarket: DataTypes.INTEGER,
+            id_anymarket: DataTypes.BIGINT,
             status_anymarket: DataTypes.STRING,
             status_bseller: DataTypes.STRING,
         }, {
